requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Database
supabase>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.31.0
httpx>=0.25.0

# Logging & Monitoring
structlog>=23.2.0

# Type Checking (Development)
mypy>=1.7.0

# LLM Clients
google-generativeai>=0.3.0
openai>=1.0.0
